backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pandas
xgboost
shap
scikit-learn
jinja2
gunicorn
uvloop
httptools
psycopg2-binary
